refactor(test): move class-progression fixture types to a types file
Extract the harness interfaces (ClassMeta, FeatureDoc, LevelSummary, etc.)
into classProgression.types.ts and drop the fixture's function re-exports, so
the fixture no longer re-exports types or source functions. Consumers now import
these types from the types file and ClassFeatureIndex from getClassFeatures
directly, fixing the type-check failures on the PR.
