+++ src/utils/classProgression/commander.test.ts
@@ -1,11 +1,11 @@
 import { afterAll, beforeAll, describe, expect, it } from 'vitest';
-import type { LevelSummary } from '../../../tests/fixtures/classProgression.ts';
 import {
 	buildRealIndex,
 	getClassMeta,
 	restoreMocks,
 	simulateProgression,
 } from '../../../tests/fixtures/classProgression.ts';
+import type { LevelSummary } from '../../../tests/fixtures/classProgression.types.ts';
 import type { ClassFeatureIndex } from '../getClassFeatures.ts';
 
 /**
