+++ src/utils/classProgression/shepherd.test.ts
@@ -1,13 +1,12 @@
 import { afterAll, beforeAll, describe, expect, it } from 'vitest';
 import {
 	buildRealIndex,
-	type ClassMeta,
 	getClassMeta,
-	type LevelSummary,
 	resolveLevel,
 	restoreMocks,
 	simulateProgression,
 } from '../../../tests/fixtures/classProgression.ts';
+import type { ClassMeta, LevelSummary } from '../../../tests/fixtures/classProgression.types.ts';
 import type { ClassFeatureIndex } from '../getClassFeatures.ts';
 
 /**
